move rearrange to torch.func
