+++ functorch/einops/__init__.py
@@ -1,4 +1,4 @@
-from .rearrange import rearrange
+from torch.func import rearrange
 
 
 __all__ = ["rearrange"]
