+++ test/functorch/test_rearrange.py
@@ -29,7 +29,7 @@ SOFTWARE.
 import numpy as np
 
 import torch
-from functorch.einops import rearrange
+from torch.func import rearrange
 from torch.testing._internal.common_utils import run_tests, TestCase
