+++ torch/func/__init__.py
@@ -10,6 +10,7 @@ from torch._functorch.eager_transforms import (
     linearize,
     vjp,
 )
+from torch._functorch.einops import rearrange
 from torch._functorch.functional_call import functional_call, stack_module_state
 
 
@@ -28,4 +29,5 @@ __all__ = [
     "functional_call",
     "stack_module_state",
     "debug_unwrap",
+    "rearrange",
 ]
